Lab_works/5lab/shop_back/requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
drf-nested-routers==0.93.4
pytz==2022.7.1
sqlparse==0.4.3
